Move patterns to a populate function and implement test pass
